Replaced PlainDisplayer with StaticDisplayer
